Centralize sampling params and test optimizations.
