+++ GPU/Common/GPUStateUtils.h
@@ -1,0 +1,115 @@
+// Copyright (c) 2012- PPSSPP Project.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, version 2.0 or later versions.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License 2.0 for more details.
+
+// A copy of the GPL 2.0 should have been included with the program.
+// If not, see http://www.gnu.org/licenses/
+
+// Official git repository and contact information can be found at
+// https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
+
+#include "GPU/ge_constants.h"
+#include "GPU/GPUState.h"
+
+// Dest factors where it's safe to eliminate the alpha test under certain conditions
+static const bool safeDestFactors[16] = {
+	true, // GE_DSTBLEND_SRCCOLOR,
+	true, // GE_DSTBLEND_INVSRCCOLOR,
+	false, // GE_DSTBLEND_SRCALPHA,
+	true, // GE_DSTBLEND_INVSRCALPHA,
+	true, // GE_DSTBLEND_DSTALPHA,
+	true, // GE_DSTBLEND_INVDSTALPHA,
+	false, // GE_DSTBLEND_DOUBLESRCALPHA,
+	false, // GE_DSTBLEND_DOUBLEINVSRCALPHA,
+	true, // GE_DSTBLEND_DOUBLEDSTALPHA,
+	true, // GE_DSTBLEND_DOUBLEINVDSTALPHA,
+	true, //GE_DSTBLEND_FIXB,
+};
+
+static inline bool IsAlphaTestTriviallyTrue() {
+	switch (gstate.getAlphaTestFunction()) {
+	case GE_COMP_NEVER:
+		return false;
+
+	case GE_COMP_ALWAYS:
+		return true;
+
+	case GE_COMP_GEQUAL:
+		if (gstate_c.vertexFullAlpha && (gstate_c.textureFullAlpha || !gstate.isTextureAlphaUsed()))
+			return true;  // If alpha is full, it doesn't matter what the ref value is.
+		return gstate.getAlphaTestRef() == 0;
+
+	// Non-zero check. If we have no depth testing (and thus no depth writing), and an alpha func that will result in no change if zero alpha, get rid of the alpha test.
+	// Speeds up Lumines by a LOT on PowerVR.
+	case GE_COMP_NOTEQUAL:
+		if (gstate.getAlphaTestRef() == 255) {
+			// Likely to be rare. Let's just skip the vertexFullAlpha optimization here instead of adding
+			// complicated code to discard the draw or whatnot.
+			return false;
+		}
+		// Fallthrough on purpose
+
+	case GE_COMP_GREATER:
+		{
+#if 0
+			// Easy way to check the values in the debugger without ruining && early-out
+			bool doTextureAlpha = gstate.isTextureAlphaUsed();
+			bool stencilTest = gstate.isStencilTestEnabled();
+			bool depthTest = gstate.isDepthTestEnabled();
+			GEComparison depthTestFunc = gstate.getDepthTestFunction();
+			int alphaRef = gstate.getAlphaTestRef();
+			int blendA = gstate.getBlendFuncA();
+			bool blendEnabled = gstate.isAlphaBlendEnabled();
+			int blendB = gstate.getBlendFuncA();
+#endif
+			return (gstate_c.vertexFullAlpha && (gstate_c.textureFullAlpha || !gstate.isTextureAlphaUsed())) || (
+					(!gstate.isStencilTestEnabled() &&
+					!gstate.isDepthTestEnabled() &&
+					gstate.getAlphaTestRef() == 0 &&
+					gstate.isAlphaBlendEnabled() &&
+					gstate.getBlendFuncA() == GE_SRCBLEND_SRCALPHA &&
+					safeDestFactors[(int)gstate.getBlendFuncB()]));
+		}
+
+	case GE_COMP_LEQUAL:
+		return gstate.getAlphaTestRef() == 255;
+
+	case GE_COMP_EQUAL:
+	case GE_COMP_LESS:
+		return false;
+
+	default:
+		return false;
+	}
+}
+
+static inline bool IsAlphaTestAgainstZero() {
+	return gstate.getAlphaTestRef() == 0 && gstate.getAlphaTestMask() == 0xFF;
+}
+
+static inline bool IsColorTestAgainstZero() {
+	return gstate.getColorTestRef() == 0 && gstate.getColorTestMask() == 0xFFFFFF;
+}
+
+static inline bool IsColorTestTriviallyTrue() {
+	switch (gstate.getColorTestFunction()) {
+	case GE_COMP_NEVER:
+		return false;
+
+	case GE_COMP_ALWAYS:
+		return true;
+
+	case GE_COMP_EQUAL:
+	case GE_COMP_NOTEQUAL:
+		return false;
+	default:
+		return false;
+	}
+}
